Move function to separate file
